Make helpers/AsyncConnect provide redux-connect exports
